working on implementing score/
just working on implementing the score object. it has a lifetime of the entire game, planning on having it contain a try score, try cross, and try end method which return booleans, and a get score which adds the totals.

if the processes of the try score or try end methods get to chunky I will potentially seperate out some pure functions.