Finished Score class for now*
finished the get total score fun. when scoreing an upper section play, if the score is >= 63 and upper bonus is false, then set upper bonus to be true. the get total score just calculates all 13 categories and then if the bonus is active. turn end control will be in the game state class.