Working on gamestate class, and updated dice
Added a method to reset the dice object for use when a turn ends. Started work on the gamestate object. This abstraction is being made with the purpose of making it easy to take this dev space working version, and then clean it up and make a web implementation with a variable number of players.