Working dev version and Test Case
Made a working response version, while the implementation is extreemly ugly and goes agains everything I strive for as a programmer. It is a 500 line Case statment of repeated code, that needs a constantly refrenced key, and the display is janky, and dosn't show total score. no input cleaning ect

This is just a logic test version to be re-implemented in WASM, and the logic was able to have a full run time doing everything you should be able to do, and the interactions that you shouldn't be able to do. And was able to apropriatly respond aditionally, It is able to vet a good range of undefined behavior with explicit function calls in a method that there arent any try catch blocks and may not need any with proper input cleaning on final version.

Added attached text file of the game instance.

--fixed mistake where the scorePlay and scoreCross functions in gameState had their names reversed.